+++ app/DataTables/Admin/Anak/ImunisasiDataTable.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace App\DataTables\Admin;
+namespace App\DataTables\admin\Anak;
 
-use App\Models\DataAnak;
+use App\Models\Imunisasi;
 use Yajra\DataTables\Html\Button;
 use Yajra\DataTables\Html\Column;
 use Yajra\DataTables\Html\Editor\Editor;
 use Yajra\DataTables\Html\Editor\Fields;
 use Yajra\DataTables\Services\DataTable;
 
-class DataAnakDataTable extends DataTable
+class ImunisasiDataTable extends DataTable
 {
     /**
      * Build DataTable class.
@@ -20,29 +20,29 @@ class DataAnakDataTable extends DataTable
     public function dataTable($query)
     {
         return datatables()
-        ->eloquent($query)
-        ->setRowId(function ($row) {
-            return $row->id;
-        })
-        ->addColumn('action', function ($row) {
-            $btn = '<div class="btn-group">';
-            $btn = $btn . '<a href="' . route('admin.dataanak.edit', $row->id) . '" class="btn btn-dark buttons-edit"><i class="fas fa-edit"></i></a>';
-            $btn = $btn . '<a href="' . route('admin.dataanak.destroy', $row->id) . '" class="btn btn-danger buttons-delete"><i class="fas fa-trash fa-fw"></i></a>';
-            $btn = $btn . '</div>';
+            ->eloquent($query)
+            ->setRowId(function ($row) {
+                return $row->id;
+            })
+            ->addColumn('action', function ($row) {
+                $btn = '<div class="btn-group">';
+                $btn = $btn . '<a href="' . route('admin.anak-data.imunisasi.edit', $row->id) . '" class="btn btn-dark buttons-edit"><i class="fas fa-edit"></i></a>';
+                $btn = $btn . '<a href="' . route('admin.anak-data.imunisasi.destroy', $row->id) . '" class="btn btn-danger buttons-delete"><i class="fas fa-trash fa-fw"></i></a>';
+                $btn = $btn . '</div>';
 
-            return $btn;
-        });
+                return $btn;
+            });
     }
 
     /**
      * Get query source of dataTable.
      *
-     * @param \App\App\Models\Admin/Master/DataAnak $model
+     * @param \App\App\Models\admin\Imunisasi $model
      * @return \Illuminate\Database\Eloquent\Builder
      */
-    public function query(DataAnak $model)
+    public function query(Imunisasi $model)
     {
-        return $model->newQuery();
+        return $model->select('imunisasis.*')->with(['jenisvaksin']);
     }
 
     /**
@@ -53,7 +53,7 @@ class DataAnakDataTable extends DataTable
     public function html()
     {
         return $this->builder()
-                    ->setTableId('dataanak-table')
+                    ->setTableId('imunisasi-table')
                     ->columns($this->getColumns())
                     ->minifiedAjax()
                     ->dom('<"dataTables_wrapper dt-bootstrap"B<"row"<"col-xl-7 d-block d-sm-flex d-xl-block justify-content-center"<"d-block d-lg-inline-flex"l>><"col-xl-5 d-flex d-xl-block justify-content-center"fr>>t<"row"<"col-sm-5"i><"col-sm-7"p>>>')
@@ -82,13 +82,14 @@ class DataAnakDataTable extends DataTable
                   ->addClass('text-center'),
             Column::make('id'),
             Column::make('nama_anak'),
-            Column::make('NIK'),
-            Column::make('tempat_lahir'),
-            Column::make('tanggal_lahir'),
-            Column::make('umur'),
             Column::make('jenis_kelamin'),
-            Column::make('anak_ke'),
-            Column::make('nama_orangtua'),
+            Column::make('tanggal_imunisasi'),
+            Column::make('berat_badan'),
+            Column::make('tinggi_badan'),
+            Column::make('umur'),
+            Column::make('jenis_vaksin')->data('jenisvaksin.vaksin'), //jenisvaksin nama fungsi relasi
+            Column::make('jadwal_vaksin'),
+            Column::make('nama_kader'),
         ];
     }
 
@@ -99,6 +100,6 @@ class DataAnakDataTable extends DataTable
      */
     protected function filename()
     {
-        return 'Admin/DataAnak_' . date('YmdHis');
+        return 'admin\Imunisasi_' . date('YmdHis');
     }
 }
