+++ resources/views/pages/admin/anak/dataanak/add-edit.blade.php
@@ -20,7 +20,7 @@
 
 
 <!-- begin panel -->
-<form action="{{ isset($data) ? route('admin.dataanak.update', $data->id) : route('admin.dataanak.store') }}" id="form" name="form" method="POST" data-parsley-validate="true">
+<form action="{{ isset($data) ? route('admin.anak-data.dataanak.update', $data->id) : route('admin.anak-data.dataanak.store') }}" id="form" name="form" method="POST" data-parsley-validate="true">
   @csrf
   @if(isset($data))
   {{ method_field('PUT') }}
