+++ resources/views/pages/admin/anak/imunisasi/add-edit.blade.php
@@ -20,7 +20,7 @@
 
 
 <!-- begin panel -->
-<form action="{{ isset($data) ? route('admin.imunisasi.update', $data->id) : route('admin.imunisasi.store') }}" id="form" name="form" method="POST" data-parsley-validate="true">
+<form action="{{ isset($data) ? route('admin.anak-data.imunisasi.update', $data->id) : route('admin.anak-data.imunisasi.store') }}" id="form" name="form" method="POST" data-parsley-validate="true">
   @csrf
   @if(isset($data))
   {{ method_field('PUT') }}
