--- resources/views/pages/admin/dataanak/coba.php
@@ -1,1 +1,0 @@
-blablabalablal
